+++ src/main/java/com/board_of_ads/configs/security/SecurityConfig.java
@@ -45,9 +45,10 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     private static final Logger logger = LoggerFactory.getLogger(SecurityConfig.class);
 
+    //Для google авторизации.
     @Qualifier("oauth2ClientContext")
     @Autowired
-    private OAuth2ClientContext oAuth2ClientContext; //Для google и vk авторизацй
+    private OAuth2ClientContext oAuth2ClientContext;
 
     @Autowired
     private AuthProvider authProvider;
@@ -153,7 +154,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
     @Bean
     public PrincipalExtractor authPrincipalExtractor() {
-        return map -> socialNetworkService.loadUserInSocialNetwork(map);
+        return map -> socialNetworkService.findOrCreateUserGoogle(map);
     }
 
     //Регистрация фильтра для перехвата авторизации до срабатывания фильтра Спринга
